Split happy_journeys_spec up into multiple files
The common behaviour at the top of happy_journeys_spec.rb has been
shifted to the JourneyHelper module and a shared context that overrides
the retrieval of the latest application data.
