+++ spec/features/happy_journeys/js/changing_do_you_work_in_a_school_from_no_to_yes_spec.rb
@@ -1,0 +1,321 @@
+require "rails_helper"
+
+RSpec.feature "Happy journeys", type: :feature do
+  include Helpers::JourneyHelper
+  include_context "retrieve latest application data"
+
+  scenario "registration journey changing do you work in a school from no to yes" do
+    visit "/"
+    expect(page).to have_text("Before you start")
+    page.click_link("Start now")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Have you already chosen an NPQ and provider?")
+    page.choose("Yes", visible: :all)
+    page.click_button("Continue")
+
+    # expect(page).to be_axe_clean
+    # TODO: aria-expanded
+    expect(page.current_path).to eql("/registration/teacher-catchment")
+    page.choose("England", visible: :all)
+    page.click_button("Continue")
+
+    expect(page.current_path).to eql("/registration/work-in-school")
+    page.choose("No", visible: :all)
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page.current_path).to eql("/registration/teacher-reference-number")
+    page.choose("Yes", visible: :all)
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page.current_path).to include("contact-details")
+    expect(page).to have_text("What's your email address?")
+    page.fill_in "What's your email address?", with: "user@example.com"
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Confirm your code")
+    expect(page).to have_text("user@example.com")
+    page.click_button("Continue")
+
+    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
+
+    expect(page).to be_axe_clean
+    page.fill_in "Enter your code", with: code
+    page.click_button("Continue")
+
+    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
+      .with(
+        headers: {
+          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
+        },
+        body: {
+          trn: "1234567",
+          date_of_birth: "1980-12-13",
+          full_name: "John Doe",
+          nino: "AB123456C",
+        },
+      )
+      .to_return(status: 200, body: participant_validator_response, headers: {})
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Check your details")
+    page.fill_in "Teacher reference number (TRN)", with: "1234567"
+    page.fill_in "Full name", with: "John Doe"
+    page.fill_in "Day", with: "13"
+    page.fill_in "Month", with: "12"
+    page.fill_in "Year", with: "1980"
+    page.fill_in "National Insurance number", with: "AB123456C"
+    page.click_button("Continue")
+
+    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Do you work in early years or childcare?")
+    page.choose("No", visible: :all)
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("What are you applying for?")
+    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Tell us about where you work")
+    page.fill_in "Name of employer", with: "Big company"
+    page.fill_in "Role", with: "Trainer"
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("How is your course being paid for?")
+    page.choose "My workplace is covering the cost", visible: :all
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Select your provider")
+    page.choose("Teach First", visible: :all)
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Sharing your NPQ information")
+    page.check("Yes, I agree my information can be shared", visible: :all)
+    page.click_button("Continue")
+
+    check_answers_page = CheckAnswersPage.new
+
+    expect(page).to be_axe_clean
+    expect(check_answers_page).to be_displayed
+
+    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
+      [summary_item.key, summary_item.value]
+    }.to_h
+
+    expect(summary_data).to eql(
+      "Full name" => "John Doe",
+      "TRN" => "1234567",
+      "Date of birth" => "13 December 1980",
+      "National Insurance number" => "AB123456C",
+      "Email" => "user@example.com",
+      "Course" => "NPQ for Senior Leadership (NPQSL)",
+      "How is your NPQ being paid for?" => "My workplace is covering the cost",
+      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
+      "Do you work in early years or childcare?" => "No",
+      "Employer" => "Big company",
+      "Lead provider" => "Teach First",
+      "Role" => "Trainer",
+      "Where do you work?" => "England",
+    )
+
+    page.click_link("Change", href: "/registration/work-in-school/change")
+
+    expect(page.current_path).to eql("/registration/work-in-school/change")
+    page.choose("Yes", visible: :all)
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page.current_path).to eql("/registration/teacher-reference-number")
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page.current_path).to include("contact-details")
+    expect(page).to have_text("What's your email address?")
+    page.click_button("Continue")
+
+    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
+      .with(
+        headers: {
+          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
+        },
+        body: {
+          trn: "1234567",
+          date_of_birth: "1980-12-13",
+          full_name: "John Doe",
+          nino: "AB123456C",
+        },
+      )
+      .to_return(status: 200, body: participant_validator_response, headers: {})
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Check your details")
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Where is your school, college or academy trust?")
+    page.fill_in "Workplace location", with: "manchester"
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Choose your workplace")
+    expect(page).to have_text("Choose from schools, trusts and 16 to 19 educational settings located in manchester")
+    within ".npq-js-reveal" do
+      page.fill_in "Enter the name of your workplace", with: "open"
+    end
+
+    expect(page).to have_content("open manchester school")
+    page.find("#school-picker__option--0").click
+    page.click_button("Continue")
+
+    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-senior-leadership")
+      .with(
+        headers: {
+          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
+        },
+      )
+      .to_return(
+        status: 200,
+        body: previously_funded_response(false),
+        headers: {
+          "Content-Type" => "application/vnd.api+json",
+        },
+      )
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("What are you applying for?")
+    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("DfE scholarship funding is not available")
+    expect(page).to have_text("To be eligible for scholarship funding for")
+    expect(page).to have_text("state-funded schools")
+    expect(page).to have_text("state-funded 16 to 19 organisations")
+    expect(page).to have_text("independent special schools")
+    expect(page).to have_text("virtual schools")
+    expect(page).to have_text("hospital schools")
+    expect(page).to have_text("young offenders institutions")
+    page.click_link("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("How is your course being paid for?")
+    page.choose "My workplace is covering the cost", visible: :all
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Select your provider")
+    page.choose("Teach First", visible: :all)
+    page.click_button("Continue")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Sharing your NPQ information")
+    page.check("Yes, I agree my information can be shared", visible: :all)
+    page.click_button("Continue")
+
+    check_answers_page = CheckAnswersPage.new
+
+    expect(page).to be_axe_clean
+    expect(check_answers_page).to be_displayed
+
+    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
+      [summary_item.key, summary_item.value]
+    }.to_h
+
+    expect(summary_data).to eql(
+      "Full name" => "John Doe",
+      "TRN" => "1234567",
+      "Date of birth" => "13 December 1980",
+      "National Insurance number" => "AB123456C",
+      "Email" => "user@example.com",
+      "Course" => "NPQ for Senior Leadership (NPQSL)",
+      "How is your NPQ being paid for?" => "My workplace is covering the cost",
+      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "Yes",
+      "Workplace" => "open manchester school",
+      "Lead provider" => "Teach First",
+      "Where do you work?" => "England",
+    )
+
+    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
+
+    page.click_button("Submit")
+
+    expect(page).to be_axe_clean
+    expect(page).to have_text("Your initial registration is complete")
+    expect(page).to_not have_text("The Early Headship Coaching Offer is a package of structured face-to-face support for new headteachers.")
+
+    expect(retrieve_latest_application_user_data).to eq(
+      "active_alert" => false,
+      "admin" => false,
+      "date_of_birth" => "1980-12-13",
+      "ecf_id" => nil,
+      "email" => "user@example.com",
+      "full_name" => "John Doe",
+      "national_insurance_number" => nil,
+      "otp_expires_at" => nil,
+      "otp_hash" => nil,
+      "trn" => "1234567",
+      "trn_auto_verified" => true,
+      "trn_verified" => true,
+    )
+    expect(retrieve_latest_application_data).to eq(
+      "cohort" => 2022,
+      "course_id" => Course.find_by_code(code: :NPQSL).id,
+      "ecf_id" => nil,
+      "eligible_for_funding" => false,
+      "employer_name" => nil,
+      "employment_role" => nil,
+      "funding_choice" => "school",
+      "funding_eligiblity_status_code" => "ineligible_establishment_type",
+      "headteacher_status" => nil,
+      "kind_of_nursery" => nil,
+      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
+      "private_childcare_provider_urn" => nil,
+      "school_urn" => "100000",
+      "targeted_delivery_funding_eligibility" => false,
+      "targeted_support_funding_eligibility" => false,
+      "teacher_catchment" => "england",
+      "teacher_catchment_country" => nil,
+      "ukprn" => nil,
+      "works_in_childcare" => false,
+      "works_in_nursery" => false,
+      "works_in_school" => true,
+      "raw_application_data" => {
+        "active_alert" => false,
+        "can_share_choices" => "1",
+        "chosen_provider" => "yes",
+        "confirmed_email" => "user@example.com",
+        "course_id" => Course.find_by_code(code: :NPQSL).id.to_s,
+        "date_of_birth" => "1980-12-13",
+        "email" => "user@example.com",
+        "employer_name" => "Big company",
+        "employment_role" => "Trainer",
+        "full_name" => "John Doe",
+        "funding" => "school",
+        "institution_identifier" => "School-100000",
+        "institution_location" => "manchester",
+        "institution_name" => "",
+        "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id.to_s,
+        "national_insurance_number" => "AB123456C",
+        "teacher_catchment" => "england",
+        "teacher_catchment_country" => nil,
+        "trn" => "1234567",
+        "trn_auto_verified" => true,
+        "trn_knowledge" => "yes",
+        "trn_verified" => true,
+        "verified_trn" => "1234567",
+        "works_in_childcare" => "no",
+        "works_in_school" => "yes",
+      },
+    )
+  end
+end
