--- spec/features/happy_journeys_spec.rb
@@ -1,3835 +1,0 @@
-require "rails_helper"
-
-RSpec.feature "Happy journeys", type: :feature do
-  def latest_application
-    Application.order(created_at: :asc).last
-  end
-
-  def latest_application_user
-    latest_application&.user
-  end
-
-  def retrieve_latest_application_user_data
-    latest_application_user&.as_json(except: %i[id created_at updated_at])
-  end
-
-  def retrieve_latest_application_data
-    latest_application&.as_json(except: %i[id created_at updated_at user_id])
-  end
-
-  before do
-    # Make sure all the tests are checking this data
-    expect(self).to receive(:retrieve_latest_application_user_data).and_call_original
-    expect(self).to receive(:retrieve_latest_application_data).and_call_original
-  end
-
-  scenario "registration journey via using old name and not headship" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("No, I need help getting one", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Get a Teacher Reference Number (TRN)")
-    page.click_link("Back")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Where is your school, college or academy trust?")
-    page.fill_in "Workplace location", with: "manchester"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Choose your workplace")
-    expect(page).to have_text("Choose from schools, trusts and 16 to 19 educational settings located in manchester")
-    within ".npq-js-reveal" do
-      page.fill_in "Enter the name of your workplace", with: "open"
-    end
-
-    expect(page).to have_content("open manchester school")
-    page.find("#school-picker__option--0").click
-    page.click_button("Continue")
-
-    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-senior-leadership")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-      )
-      .to_return(
-        status: 200,
-        body: previously_funded_response(false),
-        headers: {
-          "Content-Type" => "application/vnd.api+json",
-        },
-      )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "Workplace" => "open manchester school",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "Yes",
-      "Lead provider" => "Teach First",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Your initial registration is complete")
-    expect(page).to_not have_text("The Early Headship Coaching Offer is a package of structured face-to-face support for new headteachers.")
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQSL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "school",
-      "funding_eligiblity_status_code" => "ineligible_establishment_type",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => "100000",
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => false,
-      "works_in_school" => true,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQSL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "funding" => "school",
-        "institution_identifier" => "School-100000",
-        "institution_location" => "manchester",
-        "institution_name" => "",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_school" => "yes",
-      },
-    )
-  end
-
-  scenario "registration journey via using same name" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.fill_in "Enter your code", with: "000000"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("Code is not correct")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-    School.create!(urn: 100_001, name: "closed manchester school", address_1: "street 2", town: "manchester", establishment_status_code: "2")
-    School.create!(urn: 100_002, name: "open newcastle school", address_1: "street 3", town: "newcastle", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Where is your school, college or academy trust?")
-    page.fill_in "Workplace location", with: "manchester"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Choose your workplace")
-    expect(page).to have_text("Choose from schools, trusts and 16 to 19 educational settings located in manchester")
-    within ".npq-js-reveal" do
-      page.fill_in "Enter the name of your workplace", with: "open"
-    end
-
-    expect(page).to have_content("open manchester school")
-    page.find("#school-picker__option--0").click
-    page.click_button("Continue")
-
-    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-headship")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-      )
-      .to_return(
-        status: 200,
-        body: previously_funded_response(false),
-        headers: {
-          "Content-Type" => "application/vnd.api+json",
-        },
-      )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Headship (NPQH)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My trust is paying", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Headship (NPQH)",
-      "Lead provider" => "Teach First",
-      "Workplace" => "open manchester school",
-      "How is your NPQ being paid for?" => "My trust is paying",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "Yes",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Your initial registration is complete")
-    expect(page).to have_text("The Early Headship Coaching Offer is a package of structured face-to-face support for new headteachers.")
-
-    expect(User.count).to eql(1)
-
-    user = User.last
-
-    expect(user.email).to eql("user@example.com")
-    expect(user.full_name).to eql("John Doe")
-    expect(user.trn).to eql("1234567")
-    expect(user.trn_verified).to be_truthy
-    expect(user.trn_auto_verified).to be_truthy
-    expect(user.date_of_birth).to eql(Date.new(1980, 12, 13))
-    expect(user.national_insurance_number).to be_blank
-
-    expect(user.applications.count).to eql(1)
-
-    application = user.applications.first
-
-    expect(application.eligible_for_funding).to be_falsey
-    expect(application.funding_choice).to eql("trust")
-
-    visit "/account"
-
-    expect(page).to have_text("Teach First")
-    expect(page).to have_text("NPQ for Headship (NPQH)")
-
-    visit "/registration/share-provider"
-
-    expect(page).to have_content("Before you start")
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQH).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "trust",
-      "funding_eligiblity_status_code" => "ineligible_establishment_type",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => "100000",
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => false,
-      "works_in_school" => true,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQH).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "funding" => "trust",
-        "institution_identifier" => "School-100000",
-        "institution_location" => "manchester",
-        "institution_name" => "",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_school" => "yes",
-      },
-    )
-  end
-
-  scenario "registration journey when outside of catchment area" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("Another country", visible: :all)
-    within "[data-module='app-country-autocomplete'" do
-      page.fill_in "Which country do you teach in?", with: "Falk"
-    end
-
-    expect(page).to have_content("Falkland Islands")
-    page.find("#registration-wizard-teacher-catchment-country-field__option--0").click
-
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("No, I need help getting one", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Get a Teacher Reference Number (TRN)")
-    page.click_link("Back")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "I am paying", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "I am paying",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "Yes",
-      "Lead provider" => "Teach First",
-      "Where do you work?" => "Falkland Islands",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQSL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "self",
-      "funding_eligiblity_status_code" => "no_institution",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "another",
-      "teacher_catchment_country" => "Falkland Islands",
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => false,
-      "works_in_school" => true,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQSL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "funding" => "self",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "",
-        "teacher_catchment" => "another",
-        "teacher_catchment_country" => "Falkland Islands",
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_school" => "yes",
-      },
-    )
-  end
-
-  scenario "registration journey when outside of catchment area (crown dependencies)" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("Jersey, Guernsey or the Isle of Man", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("No, I need help getting one", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Get a Teacher Reference Number (TRN)")
-    page.click_link("Back")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "I am paying", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "I am paying",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "Yes",
-      "Lead provider" => "Teach First",
-      "Where do you work?" => "Jersey, Guernsey or the Isle of Man",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQSL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "self",
-      "funding_eligiblity_status_code" => "no_institution",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "jersey_guernsey_isle_of_man",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => false,
-      "works_in_school" => true,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQSL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "funding" => "self",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "",
-        "teacher_catchment" => "jersey_guernsey_isle_of_man",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_school" => "yes",
-      },
-    )
-  end
-
-  scenario "registration journey while not currently working at school" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Tell us about where you work")
-    page.fill_in "Name of employer", with: "Big company"
-    page.fill_in "Role", with: "Trainer"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "No",
-      "Employer" => "Big company",
-      "Lead provider" => "Teach First",
-      "Role" => "Trainer",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQSL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => "Big company",
-      "employment_role" => "Trainer",
-      "funding_choice" => "school",
-      "funding_eligiblity_status_code" => "no_institution",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => false,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => "6",
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "employer_name" => "Big company",
-        "employment_role" => "Trainer",
-        "full_name" => "John Doe",
-        "funding" => "school",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "no",
-        "works_in_school" => "no",
-      },
-    )
-  end
-
-  scenario "registration journey while working at public nursery" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in a nursery?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What kind of nursery do you work in?")
-    public_nursery_type_key = Forms::KindOfNursery::KIND_OF_NURSERY_PUBLIC_OPTIONS.sample
-    public_nursery_type = I18n.t("registration_wizard.kind_of_nursery.#{public_nursery_type_key}")
-    page.choose(public_nursery_type, visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Where is your nursery?")
-    page.fill_in "Nursery location", with: "manchester"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Choose your nursery")
-    expect(page).to have_text("Choose from nurseries located in manchester")
-    within ".npq-js-reveal" do
-      page.fill_in "Enter your nursery name", with: "open"
-    end
-
-    expect(page).to have_content("open manchester school")
-    page.find("#nursery-picker__option--0").click
-    page.click_button("Continue")
-
-    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-senior-leadership")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-      )
-      .to_return(
-        status: 200,
-        body: previously_funded_response(false),
-        headers: {
-          "Content-Type" => "application/vnd.api+json",
-        },
-      )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all) # Needs changing to an early years course once added
-    page.click_button("Continue")
-
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Do you work in a nursery?" => "Yes",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "Yes",
-      "Lead provider" => "Teach First",
-      "Nursery" => "open manchester school",
-      "Type of nursery" => public_nursery_type,
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQSL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "school",
-      "funding_eligiblity_status_code" => "ineligible_establishment_type",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => public_nursery_type_key,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => "100000",
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => true,
-      "works_in_nursery" => true,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQSL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "funding" => "school",
-        "institution_identifier" => "School-100000",
-        "institution_location" => "manchester",
-        "institution_name" => "",
-        "kind_of_nursery" => public_nursery_type_key,
-        "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id.to_s,
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "yes",
-        "works_in_nursery" => "yes",
-        "works_in_school" => "no",
-      },
-    )
-  end
-
-  scenario "registration journey while working at private nursery" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in a nursery?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What kind of nursery do you work in?")
-    page.choose("Private nursery", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you or your employer have an Ofsted unique reference number (URN)?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    private_childcare_provider = PrivateChildcareProvider.create!(
-      provider_urn: "EY123456", provider_name: "searchable childcare provider",
-      address_1: "street 1", town: "manchester",
-      early_years_individual_registers: %w[CCR VCR EYR]
-    )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Enter your or your employer's URN")
-    within ".npq-js-reveal" do
-      page.fill_in "private-childcare-provider-picker", with: "EY123"
-    end
-
-    expect(page).to have_content("EY123456 - searchable childcare provider - street 1, manchester")
-    page.find("#private-childcare-provider-picker__option--0").click
-    page.click_button("Continue")
-
-    Course::COURSE_ECF_ID_TO_IDENTIFIER_MAPPING.each_value do |course_identifier|
-      stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=#{course_identifier}")
-        .with(
-          headers: {
-            "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-          },
-        )
-        .to_return(
-          status: 200,
-          body: previously_funded_response(false),
-          headers: {
-            "Content-Type" => "application/vnd.api+json",
-          },
-        )
-    end
-
-    eyl_course = ["NPQ for Early Years Leadership (NPQEYL)"]
-
-    ineligible_courses = Forms::ChooseYourNpq.new.options.map(&:text) - eyl_course
-
-    ineligible_courses.each do |course|
-      expect(page).to have_text("What are you applying for?")
-      page.choose(course, visible: :all)
-      page.click_button("Continue")
-
-      expect(page).not_to have_text("If your provider accepts your application, you’ll qualify for DfE funding")
-      page.click_link("Back")
-    end
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Early Years Leadership (NPQEYL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("If your provider accepts your application, you’ll qualify for DfE funding")
-    expect(page).to have_text("You’ll only be eligible for DfE funding for this NPQ once. If you start this NPQ, and then withdraw or fail, you will not be funded again for the same course.")
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Course" => "NPQ for Early Years Leadership (NPQEYL)",
-      "Date of birth" => "13 December 1980",
-      "Do you work in a nursery?" => "Yes",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "Yes",
-      "Email" => "user@example.com",
-      "Full name" => "John Doe",
-      "Lead provider" => "Teach First",
-      "National Insurance number" => "AB123456C",
-      "Ofsted registration details" => private_childcare_provider.registration_details,
-      "TRN" => "1234567",
-      "Type of nursery" => "Private nursery",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQEYL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => true,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => nil,
-      "funding_eligiblity_status_code" => "funded",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => "private_nursery",
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => "EY123456",
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => true,
-      "works_in_nursery" => true,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQEYL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "has_ofsted_urn" => "yes",
-        "institution_identifier" => "PrivateChildcareProvider-EY123456",
-        "institution_name" => "",
-        "kind_of_nursery" => "private_nursery",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "yes",
-        "works_in_nursery" => "yes",
-        "works_in_school" => "no",
-      },
-    )
-  end
-
-  scenario "registration journey while working at private nursery" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in a nursery?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What kind of nursery do you work in?")
-    page.choose("Private nursery", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you or your employer have an Ofsted unique reference number (URN)?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    PrivateChildcareProvider.create!(
-      provider_urn: "EY123456", provider_name: "searchable childcare provider",
-      address_1: "street 1", town: "manchester",
-      early_years_individual_registers: %w[CCR VCR EYR]
-    )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Enter your or your employer's URN")
-    within ".npq-js-reveal" do
-      page.fill_in "private-childcare-provider-picker", with: "EY123"
-    end
-
-    expect(page).to have_content("EY123456 - searchable childcare provider - street 1, manchester")
-    page.find("#private-childcare-provider-picker__option--0").click
-    page.click_button("Continue")
-
-    Course::COURSE_ECF_ID_TO_IDENTIFIER_MAPPING.each_value do |course_identifier|
-      stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=#{course_identifier}")
-        .with(
-          headers: {
-            "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-          },
-        )
-        .to_return(
-          status: 200,
-          body: previously_funded_response(false),
-          headers: {
-            "Content-Type" => "application/vnd.api+json",
-          },
-        )
-    end
-
-    eyl_course = ["NPQ for Early Years Leadership (NPQEYL)"]
-
-    ineligible_courses = Forms::ChooseYourNpq.new.options.map(&:text) - eyl_course
-
-    ineligible_courses.each do |course|
-      expect(page).to have_text("What are you applying for?")
-      page.choose(course, visible: :all)
-      page.click_button("Continue")
-
-      expect(page).not_to have_text("If your provider accepts your application, you’ll qualify for DfE funding")
-      page.click_link("Back")
-    end
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Early Years Leadership (NPQEYL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("If your provider accepts your application, you’ll qualify for DfE funding")
-    expect(page).to have_text("You’ll only be eligible for DfE funding for this NPQ once. If you start this NPQ, and then withdraw or fail, you will not be funded again for the same course.")
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Early Years Leadership (NPQEYL)",
-      "Do you work in early years or childcare?" => "Yes",
-      "Do you work in a nursery?" => "Yes",
-      "Ofsted registration details" => "EY123456 - searchable childcare provider",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Lead provider" => "Teach First",
-      "Type of nursery" => "Private nursery",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQEYL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => true,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => nil,
-      "funding_eligiblity_status_code" => "funded",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => "private_nursery",
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => "EY123456",
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => true,
-      "works_in_nursery" => true,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQEYL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "has_ofsted_urn" => "yes",
-        "institution_identifier" => "PrivateChildcareProvider-EY123456",
-        "institution_name" => "",
-        "kind_of_nursery" => "private_nursery",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "yes",
-        "works_in_nursery" => "yes",
-        "works_in_school" => "no",
-      },
-    )
-  end
-
-  scenario "registration journey while working at private childcare provider but not a nursery" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in a nursery?")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you or your employer have an Ofsted unique reference number (URN)?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    private_childcare_provider = PrivateChildcareProvider.create!(
-      provider_urn: "EY123456",
-      provider_name: "searchable childcare provider",
-      address_1: "street 1",
-      town: "manchester",
-      early_years_individual_registers: %w[CCR VCR EYR],
-    )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Enter your or your employer's URN")
-    within ".npq-js-reveal" do
-      page.fill_in "private-childcare-provider-picker", with: "EY123"
-    end
-
-    expect(page).to have_content("EY123456 - searchable childcare provider - street 1, manchester")
-    page.find("#private-childcare-provider-picker__option--0").click
-    page.click_button("Continue")
-
-    Course::COURSE_ECF_ID_TO_IDENTIFIER_MAPPING.each_value do |course_identifier|
-      stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=#{course_identifier}")
-        .with(
-          headers: {
-            "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-          },
-        )
-        .to_return(
-          status: 200,
-          body: previously_funded_response(false),
-          headers: {
-            "Content-Type" => "application/vnd.api+json",
-          },
-        )
-    end
-
-    eyl_course = ["NPQ for Early Years Leadership (NPQEYL)"]
-
-    ineligible_courses = Forms::ChooseYourNpq.new.options.map(&:text) - eyl_course
-
-    ineligible_courses.each do |course|
-      expect(page).to have_text("What are you applying for?")
-      page.choose(course, visible: :all)
-      page.click_button("Continue")
-
-      expect(page).not_to have_text("If your provider accepts your application, you’ll qualify for DfE funding")
-      page.click_link("Back")
-    end
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Early Years Leadership (NPQEYL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("If your provider accepts your application, you’ll qualify for DfE funding")
-    expect(page).to have_text("You’ll only be eligible for DfE funding for this NPQ once. If you start this NPQ, and then withdraw or fail, you will not be funded again for the same course.")
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Course" => "NPQ for Early Years Leadership (NPQEYL)",
-      "Date of birth" => "13 December 1980",
-      "Do you work in a nursery?" => "No",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "Yes",
-      "Email" => "user@example.com",
-      "Full name" => "John Doe",
-      "Lead provider" => "Teach First",
-      "National Insurance number" => "AB123456C",
-      "Ofsted registration details" => private_childcare_provider.registration_details,
-      "TRN" => "1234567",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQEYL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => true,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => nil,
-      "funding_eligiblity_status_code" => "funded",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => "EY123456",
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => true,
-      "works_in_nursery" => false,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQEYL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "has_ofsted_urn" => "yes",
-        "institution_identifier" => "PrivateChildcareProvider-EY123456",
-        "institution_name" => "",
-        "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id.to_s,
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "yes",
-        "works_in_nursery" => "no",
-        "works_in_school" => "no",
-      },
-    )
-  end
-
-  scenario "registration journey when previously funded" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in a nursery?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What kind of nursery do you work in?")
-    page.choose("Private nursery", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you or your employer have an Ofsted unique reference number (URN)?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    PrivateChildcareProvider.create!(
-      provider_urn: "EY123456",
-      provider_name: "searchable childcare provider",
-      address_1: "street 1",
-      town: "manchester",
-      early_years_individual_registers: %w[CCR VCR EYR],
-    )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Enter your or your employer's URN")
-    within ".npq-js-reveal" do
-      page.fill_in "private-childcare-provider-picker", with: "EY123"
-    end
-
-    expect(page).to have_content("EY123456 - searchable childcare provider - street 1, manchester")
-    page.find("#private-childcare-provider-picker__option--0").click
-    page.click_button("Continue")
-
-    %w[npq-early-headship-coaching-offer npq-early-years-leadership].each do |identifier|
-      stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=#{identifier}")
-        .with(
-          headers: {
-            "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-          },
-        )
-        .to_return(
-          status: 200,
-          body: previously_funded_response(true),
-          headers: {
-            "Content-Type" => "application/vnd.api+json",
-          },
-        )
-    end
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Early Years Leadership (NPQEYL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("You can only receive scholarship funding to study this NPQ with one provider")
-    expect(page).to have_text("If you have previously failed or withdrawn from this course")
-    expect(page).to have_text("You can go back and select a different NPQ")
-    page.click_link("Back")
-
-    page.choose("Early Headship Coaching Offer", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Early Headship Coaching Offer")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Are you studying for, or have you completed an NPQ for Headship (NPQH)?")
-    page.choose("I have completed an NPQH", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Are you a headteacher?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Are you in your first 5 years of a headship?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("You can only receive scholarship funding to study this offer with one provider")
-    expect(page).to have_text("If you have previously withdrawn from this offer")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is the Early Headship Coaching Offer being paid for?")
-    page.choose "I am paying", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "Early Headship Coaching Offer",
-      "How is your EHCO being paid for?" => "I am paying",
-      "Have you completed an NPQH?" => "I have completed an NPQH",
-      "Are you a headteacher?" => "Yes",
-      "Are you in your first 5 years of a headship?" => "Yes",
-      "Do you work in a nursery?" => "Yes",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "Yes",
-      "Lead provider" => "Teach First",
-      "Ofsted registration details" => "EY123456 - searchable childcare provider",
-      "Type of nursery" => "Private nursery",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2021,
-      "course_id" => Course.find_by_code(code: :EHCO).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "self",
-      "funding_eligiblity_status_code" => "previously_funded",
-      "headteacher_status" => "yes_in_first_five_years",
-      "kind_of_nursery" => "private_nursery",
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => "EY123456",
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => true,
-      "works_in_nursery" => true,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "aso_funding_choice" => "self",
-        "aso_headteacher" => "yes",
-        "aso_new_headteacher" => "yes",
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :EHCO).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "has_ofsted_urn" => "yes",
-        "institution_identifier" => "PrivateChildcareProvider-EY123456",
-        "institution_name" => "",
-        "kind_of_nursery" => "private_nursery",
-        "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id.to_s,
-        "national_insurance_number" => "AB123456C",
-        "npqh_status" => "completed_npqh",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "yes",
-        "works_in_nursery" => "yes",
-        "works_in_school" => "no",
-      },
-    )
-  end
-
-  scenario "registration journey changing NPQ to one LeadProvider no longer supports" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in a nursery?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What kind of nursery do you work in?")
-    public_nursery_type_key = Forms::KindOfNursery::KIND_OF_NURSERY_PUBLIC_OPTIONS.sample
-    public_nursery_type = I18n.t("registration_wizard.kind_of_nursery.#{public_nursery_type_key}")
-    page.choose(public_nursery_type, visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Where is your nursery?")
-    page.fill_in "Nursery location", with: "manchester"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Choose your nursery")
-    expect(page).to have_text("Choose from nurseries located in manchester")
-    within ".npq-js-reveal" do
-      page.fill_in "Enter your nursery name", with: "open"
-    end
-
-    expect(page).to have_content("open manchester school")
-    page.find("#nursery-picker__option--0").click
-    page.click_button("Continue")
-
-    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-senior-leadership")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-      )
-      .to_return(
-        status: 200,
-        body: previously_funded_response(false),
-        headers: {
-          "Content-Type" => "application/vnd.api+json",
-        },
-      )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all) # Needs changing to an early years course once added
-    page.click_button("Continue")
-
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Best Practice Network (home of Outstanding Leaders Partnership)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Lead provider" => "Best Practice Network (home of Outstanding Leaders Partnership)",
-      "Do you work in a nursery?" => "Yes",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "Yes",
-      "Nursery" => "open manchester school",
-      "Type of nursery" => public_nursery_type,
-      "Where do you work?" => "England",
-    )
-
-    page.click_link("Change", href: "/registration/choose-your-npq/change")
-
-    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-early-years-leadership")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-      )
-      .to_return(
-        status: 200,
-        body: previously_funded_response(false),
-        headers: {
-          "Content-Type" => "application/vnd.api+json",
-        },
-      )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Early Years Leadership (NPQEYL)", visible: :all) # Needs changing to an early years course once added
-    page.click_button("Continue")
-
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    expect(page).to_not have_text("Best Practice Network (home of Outstanding Leaders Partnership)")
-    page.choose("Teacher Development Trust", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Early Years Leadership (NPQEYL)",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Lead provider" => "Teacher Development Trust",
-      "Do you work in a nursery?" => "Yes",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "Yes",
-      "Nursery" => "open manchester school",
-      "Type of nursery" => public_nursery_type,
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQEYL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "school",
-      "funding_eligiblity_status_code" => "ineligible_establishment_type",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => public_nursery_type_key,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teacher Development Trust").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => "100000",
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => true,
-      "works_in_nursery" => true,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQEYL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "funding" => "school",
-        "institution_identifier" => "School-100000",
-        "institution_location" => "manchester",
-        "institution_name" => "",
-        "kind_of_nursery" => public_nursery_type_key,
-        "lead_provider_id" => LeadProvider.find_by(name: "Teacher Development Trust").id.to_s,
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "yes",
-        "works_in_nursery" => "yes",
-        "works_in_school" => "no",
-      },
-    )
-  end
-
-  scenario "registration journey changing do you work in a school from no to yes" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Tell us about where you work")
-    page.fill_in "Name of employer", with: "Big company"
-    page.fill_in "Role", with: "Trainer"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "No",
-      "Employer" => "Big company",
-      "Lead provider" => "Teach First",
-      "Role" => "Trainer",
-      "Where do you work?" => "England",
-    )
-
-    page.click_link("Change", href: "/registration/work-in-school/change")
-
-    expect(page.current_path).to eql("/registration/work-in-school/change")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Where is your school, college or academy trust?")
-    page.fill_in "Workplace location", with: "manchester"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Choose your workplace")
-    expect(page).to have_text("Choose from schools, trusts and 16 to 19 educational settings located in manchester")
-    within ".npq-js-reveal" do
-      page.fill_in "Enter the name of your workplace", with: "open"
-    end
-
-    expect(page).to have_content("open manchester school")
-    page.find("#school-picker__option--0").click
-    page.click_button("Continue")
-
-    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-senior-leadership")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-      )
-      .to_return(
-        status: 200,
-        body: previously_funded_response(false),
-        headers: {
-          "Content-Type" => "application/vnd.api+json",
-        },
-      )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "Yes",
-      "Workplace" => "open manchester school",
-      "Lead provider" => "Teach First",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Your initial registration is complete")
-    expect(page).to_not have_text("The Early Headship Coaching Offer is a package of structured face-to-face support for new headteachers.")
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQSL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "school",
-      "funding_eligiblity_status_code" => "ineligible_establishment_type",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => "100000",
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => false,
-      "works_in_school" => true,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQSL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "employer_name" => "Big company",
-        "employment_role" => "Trainer",
-        "full_name" => "John Doe",
-        "funding" => "school",
-        "institution_identifier" => "School-100000",
-        "institution_location" => "manchester",
-        "institution_name" => "",
-        "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id.to_s,
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "no",
-        "works_in_school" => "yes",
-      },
-    )
-  end
-
-  scenario "registration journey changing do you work in childcare from yes to no" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in a nursery?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What kind of nursery do you work in?")
-    public_nursery_type_key = Forms::KindOfNursery::KIND_OF_NURSERY_PUBLIC_OPTIONS.sample
-    public_nursery_type = I18n.t("registration_wizard.kind_of_nursery.#{public_nursery_type_key}")
-    page.choose(public_nursery_type, visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Where is your nursery?")
-    page.fill_in "Nursery location", with: "manchester"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Choose your nursery")
-    expect(page).to have_text("Choose from nurseries located in manchester")
-    within ".npq-js-reveal" do
-      page.fill_in "Enter your nursery name", with: "open"
-    end
-
-    expect(page).to have_content("open manchester school")
-    page.find("#nursery-picker__option--0").click
-    page.click_button("Continue")
-
-    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-senior-leadership")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-      )
-      .to_return(
-        status: 200,
-        body: previously_funded_response(false),
-        headers: {
-          "Content-Type" => "application/vnd.api+json",
-        },
-      )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "Yes",
-      "Do you work in a nursery?" => "Yes",
-      "Type of nursery" => public_nursery_type,
-      "Lead provider" => "Teach First",
-      "Nursery" => "open manchester school",
-      "Where do you work?" => "England",
-    )
-
-    page.click_link("Change", href: "/registration/work-in-childcare/change")
-
-    expect(page.current_path).to eql("/registration/work-in-childcare/change")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Tell us about where you work")
-    page.fill_in "Name of employer", with: "Big company"
-    page.fill_in "Role", with: "Trainer"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "I am paying", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "Employer" => "Big company",
-      "Role" => "Trainer",
-      "How is your NPQ being paid for?" => "I am paying",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "No",
-      "Lead provider" => "Teach First",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQSL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => "Big company",
-      "employment_role" => "Trainer",
-      "funding_choice" => "self",
-      "funding_eligiblity_status_code" => "no_institution",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => public_nursery_type_key,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => true,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQSL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "employer_name" => "Big company",
-        "employment_role" => "Trainer",
-        "full_name" => "John Doe",
-        "funding" => "self",
-        "institution_identifier" => "School-100000",
-        "institution_location" => "manchester",
-        "institution_name" => "",
-        "kind_of_nursery" => public_nursery_type_key,
-        "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id.to_s,
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "no",
-        "works_in_nursery" => "yes",
-        "works_in_school" => "no",
-      },
-    )
-  end
-
-  scenario "registration journey changing from outside of catchment area to inside" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("Another country", visible: :all)
-    within "[data-module='app-country-autocomplete'" do
-      page.fill_in "Which country do you teach in?", with: "Falk"
-    end
-
-    expect(page).to have_content("Falkland Islands")
-    page.find("#registration-wizard-teacher-catchment-country-field__option--0").click
-
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("No, I need help getting one", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Get a Teacher Reference Number (TRN)")
-    page.click_link("Back")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "I am paying", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "How is your NPQ being paid for?" => "I am paying",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "Yes",
-      "Lead provider" => "Teach First",
-      "Where do you work?" => "Falkland Islands",
-    )
-
-    page.click_link("Change", href: "/registration/teacher-catchment/change")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment/change")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Where is your school, college or academy trust?")
-    page.fill_in "Workplace location", with: "manchester"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Choose your workplace")
-    expect(page).to have_text("Choose from schools, trusts and 16 to 19 educational settings located in manchester")
-    within ".npq-js-reveal" do
-      page.fill_in "Enter the name of your workplace", with: "open"
-    end
-
-    expect(page).to have_content("open manchester school")
-    page.find("#school-picker__option--0").click
-    page.click_button("Continue")
-
-    stub_request(:get, "https://ecf-app.gov.uk/api/v1/npq-funding/1234567?npq_course_identifier=npq-senior-leadership")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-      )
-      .to_return(
-        status: 200,
-        body: previously_funded_response(false),
-        headers: {
-          "Content-Type" => "application/vnd.api+json",
-        },
-      )
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Senior Leadership (NPQSL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("DfE scholarship funding is not available")
-    expect(page).to have_text("To be eligible for scholarship funding for")
-    expect(page).to have_text("state-funded schools")
-    expect(page).to have_text("state-funded 16 to 19 organisations")
-    expect(page).to have_text("independent special schools")
-    expect(page).to have_text("virtual schools")
-    expect(page).to have_text("hospital schools")
-    expect(page).to have_text("young offenders institutions")
-    page.click_link("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "My workplace is covering the cost", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Senior Leadership (NPQSL)",
-      "Workplace" => "open manchester school",
-      "How is your NPQ being paid for?" => "My workplace is covering the cost",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "Yes",
-      "Lead provider" => "Teach First",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Your initial registration is complete")
-    expect(page).to_not have_text("The Early Headship Coaching Offer is a package of structured face-to-face support for new headteachers.")
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQSL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => nil,
-      "employment_role" => nil,
-      "funding_choice" => "school",
-      "funding_eligiblity_status_code" => "ineligible_establishment_type",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => "100000",
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => false,
-      "works_in_school" => true,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => Course.find_by_code(code: :NPQSL).id.to_s,
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "full_name" => "John Doe",
-        "funding" => "school",
-        "institution_identifier" => "School-100000",
-        "institution_location" => "manchester",
-        "institution_name" => "",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_school" => "yes",
-      },
-    )
-  end
-
-  scenario "registration journey while working in neither a school nor childcare" do
-    visit "/"
-    expect(page).to have_text("Before you start")
-    page.click_link("Start now")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Have you already chosen an NPQ and provider?")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    # expect(page).to be_axe_clean
-    # TODO: aria-expanded
-    expect(page.current_path).to eql("/registration/teacher-catchment")
-    page.choose("England", visible: :all)
-    page.click_button("Continue")
-
-    expect(page.current_path).to eql("/registration/work-in-school")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to eql("/registration/teacher-reference-number")
-    page.choose("Yes", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page.current_path).to include("contact-details")
-    expect(page).to have_text("What's your email address?")
-    page.fill_in "What's your email address?", with: "user@example.com"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Confirm your code")
-    expect(page).to have_text("user@example.com")
-    page.click_button("Continue")
-
-    code = ActionMailer::Base.deliveries.last[:personalisation].unparsed_value[:code]
-
-    expect(page).to be_axe_clean
-    page.fill_in "Enter your code", with: code
-    page.click_button("Continue")
-
-    stub_request(:post, "https://ecf-app.gov.uk/api/v1/participant-validation")
-      .with(
-        headers: {
-          "Authorization" => "Bearer ECFAPPBEARERTOKEN",
-        },
-        body: {
-          trn: "1234567",
-          date_of_birth: "1980-12-13",
-          full_name: "John Doe",
-          nino: "AB123456C",
-        },
-      )
-      .to_return(status: 200, body: participant_validator_response, headers: {})
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Check your details")
-    page.fill_in "Teacher reference number (TRN)", with: "1234567"
-    page.fill_in "Full name", with: "John Doe"
-    page.fill_in "Day", with: "13"
-    page.fill_in "Month", with: "12"
-    page.fill_in "Year", with: "1980"
-    page.fill_in "National Insurance number", with: "AB123456C"
-    page.click_button("Continue")
-
-    School.create!(urn: 100_000, name: "open manchester school", address_1: "street 1", town: "manchester", establishment_status_code: "1")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Do you work in early years or childcare?")
-    page.choose("No", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("What are you applying for?")
-    page.choose("NPQ for Early Years Leadership (NPQEYL)", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Tell us about where you work")
-    page.fill_in "Name of employer", with: "Big company"
-    page.fill_in "Role", with: "Trainer"
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("How is your course being paid for?")
-    page.choose "I am paying", visible: :all
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Select your provider")
-    page.choose("Teach First", visible: :all)
-    page.click_button("Continue")
-
-    expect(page).to be_axe_clean
-    expect(page).to have_text("Sharing your NPQ information")
-    page.check("Yes, I agree my information can be shared", visible: :all)
-    page.click_button("Continue")
-
-    check_answers_page = CheckAnswersPage.new
-
-    expect(page).to be_axe_clean
-    expect(check_answers_page).to be_displayed
-
-    summary_data = check_answers_page.summary_list.rows.map { |summary_item|
-      [summary_item.key, summary_item.value]
-    }.to_h
-
-    expect(summary_data).to eql(
-      "Full name" => "John Doe",
-      "TRN" => "1234567",
-      "Date of birth" => "13 December 1980",
-      "National Insurance number" => "AB123456C",
-      "Email" => "user@example.com",
-      "Course" => "NPQ for Early Years Leadership (NPQEYL)",
-      "Employer" => "Big company",
-      "Role" => "Trainer",
-      "How is your NPQ being paid for?" => "I am paying",
-      "Do you work in a school, academy trust, or 16 to 19 educational setting?" => "No",
-      "Do you work in early years or childcare?" => "No",
-      "Lead provider" => "Teach First",
-      "Where do you work?" => "England",
-    )
-
-    allow(ApplicationSubmissionJob).to receive(:perform_later).with(anything)
-
-    page.click_button("Submit")
-
-    expect(page).to be_axe_clean
-
-    expect(retrieve_latest_application_user_data).to eq(
-      "active_alert" => false,
-      "admin" => false,
-      "date_of_birth" => "1980-12-13",
-      "ecf_id" => nil,
-      "email" => "user@example.com",
-      "full_name" => "John Doe",
-      "national_insurance_number" => nil,
-      "otp_expires_at" => nil,
-      "otp_hash" => nil,
-      "trn" => "1234567",
-      "trn_auto_verified" => true,
-      "trn_verified" => true,
-    )
-    expect(retrieve_latest_application_data).to eq(
-      "cohort" => 2022,
-      "course_id" => Course.find_by_code(code: :NPQEYL).id,
-      "ecf_id" => nil,
-      "eligible_for_funding" => false,
-      "employer_name" => "Big company",
-      "employment_role" => "Trainer",
-      "funding_choice" => "self",
-      "funding_eligiblity_status_code" => "no_institution",
-      "headteacher_status" => nil,
-      "kind_of_nursery" => nil,
-      "lead_provider_id" => LeadProvider.find_by(name: "Teach First").id,
-      "private_childcare_provider_urn" => nil,
-      "school_urn" => nil,
-      "targeted_delivery_funding_eligibility" => false,
-      "targeted_support_funding_eligibility" => false,
-      "teacher_catchment" => "england",
-      "teacher_catchment_country" => nil,
-      "ukprn" => nil,
-      "works_in_childcare" => false,
-      "works_in_nursery" => false,
-      "works_in_school" => false,
-      "raw_application_data" => {
-        "active_alert" => false,
-        "can_share_choices" => "1",
-        "chosen_provider" => "yes",
-        "confirmed_email" => "user@example.com",
-        "course_id" => "9",
-        "date_of_birth" => "1980-12-13",
-        "email" => "user@example.com",
-        "employer_name" => "Big company",
-        "employment_role" => "Trainer",
-        "full_name" => "John Doe",
-        "funding" => "self",
-        "lead_provider_id" => "9",
-        "national_insurance_number" => "AB123456C",
-        "teacher_catchment" => "england",
-        "teacher_catchment_country" => nil,
-        "trn" => "1234567",
-        "trn_auto_verified" => true,
-        "trn_knowledge" => "yes",
-        "trn_verified" => true,
-        "verified_trn" => "1234567",
-        "works_in_childcare" => "no",
-        "works_in_school" => "no",
-      },
-    )
-  end
-end
