+++ spec/spec_helper.rb
@@ -3,6 +3,7 @@ SimpleCov.start "rails"
 
 require "webmock/rspec"
 require "with_model"
+require "support/helpers/journey_helper"
 
 WebMock.disable_net_connect!(
   allow_localhost: true,
